QA MERGE 10037 and 10065 ***NO_CI***

+++ Dev/Dev2.Data/Decisions/Operations/IsLessThan.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 
 namespace Dev2.Data.Decisions.Operations
 {
@@ -12,14 +11,9 @@ namespace Dev2.Data.Decisions.Operations
 
         public bool Invoke(string[] cols)
         {
-            //if (cols.Length < 2 || cols.Length > 2)
-            //{
-            //    throw new InvalidDataException("Wrong number of columns sent");
-            //}
-
             if(!string.IsNullOrEmpty(cols[0]))
             {
-                int[] tryGetNumber;
+                decimal[] tryGetNumber;
                 var isString = DecisionUtils.IsNumericComparison(cols, out tryGetNumber);
 
                 //either int compare
